fix: rename MainScreen -> NeuroScreen
