+++ app/src/main/java/com/neurosky/zonetrainer/ui/neuro/NeuroScreen.kt
@@ -1,10 +1,5 @@
 package com.neurosky.zonetrainer.ui.neuro
 
-import android.content.Context
-import androidx.camera.core.CameraSelector
-import androidx.camera.core.Preview
-import androidx.camera.lifecycle.ProcessCameraProvider
-import androidx.camera.view.PreviewView
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -13,29 +8,21 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.content.ContextCompat
-import kotlin.coroutines.resume
-import kotlin.coroutines.suspendCoroutine
 
 @Composable
-fun MainScreen(viewModel: NeuroViewModel) {
+fun NeuroScreen(viewModel: NeuroViewModel) {
     val uiState by viewModel.uiState.collectAsState()
 
-    MainScreen(uiState = uiState)
+    NeuroScreen(uiState = uiState)
 }
 
 @Composable
-fun MainScreen(
+fun NeuroScreen(
     uiState: NeuroUiState
 ) {
     Box(
@@ -45,7 +32,10 @@ fun MainScreen(
         when (uiState) {
             is NeuroUiState.Connected -> {
                 if (uiState != NeuroUiState.Connected.INIT) {
-                    MainContent(attention = uiState.attention, meditation = uiState.meditation)
+                    NeuroContent(
+                        attention = uiState.attention,
+                        meditation = uiState.meditation
+                    )
                 } else {
                     Loading()
                 }
@@ -59,57 +49,6 @@ fun MainScreen(
         }
     }
 }
-
-@Composable
-fun MainContent(
-    attention: Int,
-    meditation: Int
-) {
-    val lensFacing = CameraSelector.LENS_FACING_BACK
-    val context = LocalContext.current
-    val lifecycleOwner = LocalLifecycleOwner.current
-
-    val preview = Preview.Builder().build()
-    val previewView = remember { PreviewView(context) }
-    val cameraSelector = CameraSelector.Builder()
-        .requireLensFacing(lensFacing)
-        .build()
-
-    LaunchedEffect(lensFacing) {
-        val cameraProvider = context.getCameraProvider()
-        cameraProvider.unbindAll()
-        cameraProvider.bindToLifecycle(
-            lifecycleOwner,
-            cameraSelector,
-            preview
-        )
-
-        preview.setSurfaceProvider(previewView.surfaceProvider)
-    }
-
-
-    Box(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        AndroidView({ previewView }, modifier = Modifier.fillMaxSize())
-
-        Column(
-            modifier = Modifier.align(Alignment.Center)
-        ) {
-            Text(text = "Attention: $attention")
-            Text(text = "Meditation: $meditation")
-        }
-    }
-}
-
-private suspend fun Context.getCameraProvider(): ProcessCameraProvider =
-    suspendCoroutine { continuation ->
-        ProcessCameraProvider.getInstance(this).also { cameraProvider ->
-            cameraProvider.addListener({
-                continuation.resume(cameraProvider.get())
-            }, ContextCompat.getMainExecutor(this))
-        }
-    }
 
 @Composable
 fun Loading() {
